% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tracks.R
\name{load_tracks}
\alias{load_tracks}
\title{Import tracklog data}
\usage{
load_tracks(x)
}
\arguments{
\item{x}{Name of .gpx file location}
}
\value{
Function \code{load_tracks} generates a
data.frame with points from .gpx tracklogs.
}
\description{
Imports .gpx file prior to cleaning to human readable
values in a data.frame.
}
\examples{
\dontrun{
}
}

