% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_waypoints.R
\name{load_waypoints}
\alias{load_waypoints}
\title{Import waypoints from .gpx file.}
\usage{
load_waypoints(x)
}
\arguments{
\item{x}{Name of .gpx file location}
}
\value{
Function \code{load_waypoints} generates a
data.frame with points from .gpx waypoints.
}
\description{
Imports .gpx file prior to cleaning to human readable
values in a data.frame.
}
\examples{
\dontrun{
}
}
