% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks.R
\name{make_tracks}
\alias{make_tracks}
\title{Import and summarise tracklogs and waypoints.}
\usage{
make_tracks(folder_gpx, river_shape, folder_result, date_start = NA,
  date_end = NA, make_shape = FALSE, do_waypoints = FALSE,
  make_shape_wp = FALSE, type_csv = NA)
}
\arguments{
\item{folder_gpx}{Location of folder containing .gpx files.}

\item{river_shape}{Location of file containing shapefile to define survey zones}

\item{folder_result}{Location where results are saved.}

\item{date_start}{Optional. Character of start date ("dd/mm/yyyy")}

\item{date_end}{Optional. Character of end date ("dd/mm/yyyy")}

\item{make_shape}{Optional. Logical (TRUE/FALSE). Write a shapefile with all tracklog points.}

\item{do_waypoints}{Optional. Logical (TRUE/FALSE). Generate .csv with waypoints.}

\item{make_shape_wp}{Optional. Logical (TRUE/FALSE). Generate shapefile with waypoints.}

\item{type_csv}{Required. What type of .csv to export? ".csv" or ".csv2"}
}
\value{
Will generate R objects and export .csv and shapefiles.
}
\description{
Import and summarise tracklogs and waypoints.
}
\examples{
\dontrun{
# 1) need to specify 3 locations.
#1.1) Location of shapefile with river zones
#1.2) Location of folder with .gpx files censu
#1.3) Location of folder for results

rs <- system.file("shapes/riverzones.shp", package = "rgps")
fg <- system.file("gpx", package = "rgps")
fr <- "/Location/of/result/folder"

# 2) run
list_track_data <- rgps::make_tracks(
  folder_gpx = fg, folder_result = fr,  river_shape = rs,
  do_waypoints = TRUE, make_shape = FALSE,
  make_shape_wp = FALSE,
  type_csv = ".csv")
}
}
