% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_waypoints.R
\name{make_waypoints}
\alias{make_waypoints}
\title{Import and summarise waypoints.}
\usage{
make_waypoints(folder_gpx, river_shape, folder_result, track_data = NA,
  date_start = NA, date_end = NA, make_shape_wp = FALSE)
}
\arguments{
\item{folder_gpx}{Location of folder containing .gpx files.}

\item{river_shape}{Location of folder containing shapefile to define survey zones}

\item{folder_result}{Location where results are saved.}

\item{track_data}{Optional. List with track data.frame in function \code{make_tracks}.
Used to include trip details.}

\item{date_start}{Optional. Character of start date ("dd/mm/yyyy").}

\item{date_end}{Character of end date ("dd/mm/yyyy").}

\item{make_shape_wp}{Logical (TRUE/FALSE). Generate shapefile with waypoints.}
}
\value{
Writes a .csv file and shapefiles with waypoints from .gpx files.
}
\description{
Import waypoints from .gpx files and convert to
 .csv and shapefiels.
}
\examples{
\dontrun{
}
}
