% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks.R
\name{make_tracks}
\alias{make_tracks}
\title{Import and summarise tracklogs and waypoints.}
\usage{
make_tracks(folder_gpx, river_shape, folder_result, date_start = NA,
  date_end = NA, make_shape = FALSE, do_waypoints = FALSE,
  make_shape_wp = FALSE)
}
\arguments{
\item{folder_gpx}{Location of folder containing .gpx files.}

\item{river_shape}{Location of folder containing shapefile to define survey zones}

\item{folder_result}{Location where results are saved.}

\item{date_start}{Optional. Character of start date ("dd/mm/yyyy")}

\item{date_end}{Optional. Character of end date ("dd/mm/yyyy")}

\item{make_shape}{Optional. Logical (TRUE/FALSE). Write a shapefile with all tracklog points.}

\item{do_waypoints}{Optional. Logical (TRUE/FALSE). Generate .csv with waypoints.}

\item{make_shape_wp}{Optional. Logical (TRUE/FALSE). Generate shapefile with waypoints.}
}
\value{
Will generate R objects and export .csv and shapefiles.
}
\description{
Import and summarise tracklogs and waypoints.
}
\examples{
\dontrun{
}
}

